app/requirements.txt
Flask==2.2.2
mysql-connector-python==8.0.28
flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.45
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.0

# alembic==1.9.0
# bcrypt==4.0.1
# click==8.1.3
# greenlet==2.0.1
# itsdangerous==2.1.2
# Jinja2==3.1.2
# Mako==1.2.4
# MarkupSafe==2.1.1
# mysqlclient==2.1.1
# Werkzeug==2.2.2
